Changed an argument to the appropriate type

+++ src/main/java/net/niccossystem/perfectarena/arena/PvPArena.java
@@ -64,7 +64,7 @@ public class PvPArena {
      */
     public void removePlayer(Player player) {
         players.remove(player);
-        player.damage(Double.MAX_VALUE);
+        player.damage(Integer.MAX_VALUE);
     }
 
     public Location getRandomSpawnPoint() {
